refactor(phase7): cleanup and final optimization
- Removed unused imports and duplicate code
- Moved validation functions to utils/validation.ts
- Removed repository constants (now in update module)
- Fixed missing imports (ErrorCode, SECURITY_LIMITS)
- Updated refactoring progress documentation with final results
- Final result: index.ts reduced from 3,000+ to 1,314 lines (56% reduction)
