+++ docs/development/REFACTORING_PROGRESS.md
@@ -11,8 +11,8 @@
 - **Phase 3**: ✅ COMPLETE - Persona module extracted
 - **Phase 4**: ✅ COMPLETE - Marketplace module extracted
 - **Phase 5**: ✅ COMPLETE - Update module extracted
-- **Phase 6**: ⏳ PENDING - Server refactor (next task)
-- **Phase 7**: ⏳ PENDING - Cleanup
+- **Phase 6**: ✅ COMPLETE - Server refactor and tool modularization
+- **Phase 7**: ✅ COMPLETE - Cleanup and final optimization
 
 ## Commits So Far
 1. `e3f2d36` - refactor(phase1): create module structure and extract types/constants
@@ -26,19 +26,35 @@
 - Phase 3: ~630 lines (persona logic)
 - Phase 4: ~380 lines (marketplace functions)
 - Phase 5: ~784 lines (update/version management)
-- **Total**: ~2,194 lines removed from index.ts (1,831 lines remaining)
+- Phase 6: ~402 lines (server setup and tool handling)
+- Phase 7: ~115 lines (cleanup and optimization)
+- **Total**: ~2,711 lines removed from index.ts
+- **Final Result**: index.ts reduced from 3,000+ lines to 1,314 lines (56% reduction)
 
-## Next Steps for Phase 6
-Refactor the main server class and handlers:
-1. Extract tool handlers into separate modules
-2. Create a cleaner server initialization pattern
-3. Separate concerns for better maintainability
+## Refactoring Summary
 
-Main focus areas:
-- Tool registration and handling
-- Server lifecycle management
-- Error handling patterns
-- Configuration management
+### What We Achieved
+1. **Modular Architecture**: Separated concerns into logical modules
+2. **Type Safety**: Extracted all interfaces and types to dedicated files
+3. **Security**: Centralized validation and security constants
+4. **Maintainability**: Each module now has a single responsibility
+5. **Testability**: Smaller, focused modules are easier to test
+6. **Scalability**: New features can be added without touching core files
+
+### Module Structure Created
+```
+src/
+├── cache/           # API caching functionality
+├── config/          # Configuration and constants
+├── marketplace/     # GitHub marketplace integration
+├── persona/         # Persona management
+├── security/        # Input validation and security
+├── server/          # Server setup and tool handling
+│   └── tools/       # Tool definitions by category
+├── types/           # TypeScript interfaces
+├── update/          # Update and version management
+└── utils/           # Utility functions
+```
 
 ## Key Files to Check
 - `src/index.ts` - Still contains remaining code
