+++ src/utils/validation.ts
@@ -1,0 +1,43 @@
+/**
+ * Additional validation utilities
+ */
+
+// Input validation patterns
+export const VALIDATION_PATTERNS = {
+  SAFE_FILENAME: /^[a-zA-Z0-9][a-zA-Z0-9\-_.]{0,250}[a-zA-Z0-9]$/,
+  SAFE_PATH: /^[a-zA-Z0-9\/\-_.]{1,500}$/,
+  SAFE_USERNAME: /^[a-zA-Z0-9][a-zA-Z0-9\-_.]{0,30}[a-zA-Z0-9]$/,
+  SAFE_CATEGORY: /^[a-zA-Z][a-zA-Z0-9\-_]{0,20}$/,
+  SAFE_EMAIL: /^[^\s@]+@[^\s@]+\.[^\s@]+$/
+};
+
+export function validateUsername(username: string): string {
+  if (!username || typeof username !== 'string') {
+    throw new Error('Username must be a non-empty string');
+  }
+  
+  if (!VALIDATION_PATTERNS.SAFE_USERNAME.test(username)) {
+    throw new Error('Invalid username format. Use alphanumeric characters, hyphens, underscores, and dots only.');
+  }
+  
+  return username.toLowerCase();
+}
+
+export function validateCategory(category: string): string {
+  if (!category || typeof category !== 'string') {
+    throw new Error('Category must be a non-empty string');
+  }
+  
+  if (!VALIDATION_PATTERNS.SAFE_CATEGORY.test(category)) {
+    throw new Error('Invalid category format. Use alphabetic characters, hyphens, and underscores only.');
+  }
+  
+  const validCategories = ['creative', 'professional', 'educational', 'gaming', 'personal'];
+  const normalized = category.toLowerCase();
+  
+  if (!validCategories.includes(normalized)) {
+    throw new Error(`Invalid category. Must be one of: ${validCategories.join(', ')}`);
+  }
+  
+  return normalized;
+}
